Remove GEM
For the sake of easier PR review
