--- Gemfile
@@ -1,6 +1,0 @@
-source "https://rubygems.org"
-
-git_source(:github) {|repo_name| "https://github.com/#{repo_name}" }
-
-# Specify your gem's dependencies in maxwell-interview.gemspec
-gemspec
