+++ README.md
@@ -1,63 +1,52 @@
-# Score Card Feature
+## Repository for Maxwell Interview Exercises
 
-Adds Ability to Score a Github User
+### Tips and best practices
 
-## Why Gem?
+#### 1. Send pull request
 
-In the spirit of consistency and portability, I wrapped the exercises with a gem. Makes for easy and familiar testing and dependency management
+Please clone the repo and send a pull requests. Do not push to `master` directly.
 
-## Setup
+#### 2. Branch name
 
-```
-git clone, or git pull
-git checkout ex-1-github-score-challenge
-bundle
-```
+When you work on any exercise, choose a branch name that you think would make sense to you and the reader of the pull request.
 
-## Run Test Suite
+#### 3. How to approach the problem
 
-```
-bundle exec rake test
-```
+Approach this problem as if it is an application going to production.
+We don't expect it to be perfect (no production code is), but we also don't want you to hack together a throw-away script.
+This should be representative of something that you would be comfortable releasing to a production environment.
 
-Test output
+Also, spend whatever amount of time you think is reasonable.
+You may use whatever libraries, frameworks and tools that you think are appropriate, just provide any special setup instructions when you submit your solution.
 
-```
-...S...............
-Fabulous run in 0.006579s, 2887.9769 runs/s, 2735.9781 assertions/s.
-19 runs, 18 assertions, 0 failures, 0 errors, 1 skips
-```
+You will not get any extra point for solving the challenge fast. Speed of delivery is not what we are looking for in this challenge.
 
-### Skipped Test?
+We are looking for you to demonstrate your knowledge related to common software practices to include reusability, portability and encapsulation - to name a few.
 
-The skipped test is a real call to the Github API. It's a smoke test to ensure things actually work. Uncomment to make a real HTTP call.
+#### 4. Do not fork the question
 
-## Class Summary
+Please do not fork this repo.
 
-There are three classes: `Base` API, `GithubUsers` API, and a `ScoreCard`
+#### 5. No extra points for writing tests for the exercises
 
-### Score Card
+If you want to write test for your solutions then please go ahead, but no points will be added for having tests, nor will points be deducted for not having any tests.
 
-Provides a ScoreCard class which gets data from a github user and scores their activity based on an established point system. By default, the user is 'dhh' but any user can be scored by doing the following;
+#### 6. Make commits incrementally
 
-```
-  user = 'davewoodall'
-  score_card = Maxwell::ScoreCard.new(user: user)
-  score_card.score
-  => 109
-```
+Please do not squash commit before submission. We want to see incremental commits.
 
-### GithubUser API
+#### 7. Commits after feedback should be incremental
 
-Example GET request body
+- Once you submit the answer you might get some feedback.
+- If you change code based on feedback then please make those changes as separate commit.
+- As mentioned above always make incremental commits.
 
-```
-  user = 'davewoodall'
-  request = Maxwell::Http::GithubUsers.new(user: user)
-  req = request.get
-  req.body
-```
+#### 8. There is no time limit and no extra point for doing it fast
 
-### Base API
+Most of the people taking the challenges already have a full time job.
+So take your time. Don't rush through it. There is no extra point
+for doing it fast.
 
-Create a generic HTTP layer to encapsulate use of HTTParty
+#### 9. Programming Language
+
+Your choice on programming language to use, however the Maxwell stack currently is Ruby and Javascript.
