+++ exercise_02.md
@@ -1,7 +1,7 @@
 ## Exercise 2: Price Calculator for a local grocery store
 
-At the local grocery store items are sold by quantity. For example Milk costs $3.97. 
-However sometimes there is a sale and then consumers can buy `n` items for the price of one. 
+At the local grocery store items are sold by quantity. For example Milk costs \$3.97.
+However sometimes there is a sale and then consumers can buy `n` items for the price of one.
 
 This week the pricing table at the local grocery store looks like this:
 
@@ -14,7 +14,7 @@ Banana    $0.99
 Apple     $0.89
 ```
 
-At the cash counter the items are accepted in any order. So at the cash counter the order of items can be in the order of Bread, Banana, Milk, Apple, Bread and then Milk. Notice that quantity of the Milk is 2 and this week there is a sale on milk if 2 bottles of milk are purchased. So the price applied for milk should be $5.00.
+At the cash counter the items are accepted in any order. So at the cash counter the order of items can be in the order of Bread, Banana, Milk, Apple, Bread and then Milk. Notice that quantity of the Milk is 2 and this week there is a sale on milk if 2 bottles of milk are purchased. So the price applied for milk should be \$5.00.
 
 Also note that if the user buys three bottles of milk then for the first two bottles of milk the user should get the sale price and on the third bottle of milk the user should get the unit price.
 
@@ -32,7 +32,7 @@ Item     Quantity      Price
 Milk      3            $8.97
 Bread     4            $8.17
 Apple     1            $0.89
-Banana    1            $0.99  
+Banana    1            $0.99
 
 Total price : $19.02
 You saved $3.45 today.
