--- lib/maxwell/version.rb
@@ -1,3 +1,0 @@
-module Maxwell
-  VERSION = "0.1.0"
-end
