+++ test/test_helper.rb
@@ -1,5 +1,4 @@
 $LOAD_PATH.unshift File.expand_path("../../lib", __FILE__)
-require "maxwell"
 
 require "minitest/autorun"
 require "minitest/pride"
