--- todo/README.md
@@ -1,52 +1,0 @@
-## Repository for Maxwell Interview Exercises
-
-### Tips and best practices
-
-#### 1. Send pull request
-
-Please clone the repo and send a pull requests. Do not push to `master` directly.
-
-#### 2. Branch name
-
-When you work on any exercise, choose a branch name that you think would make sense to you and the reader of the pull request.
-
-#### 3. How to approach the problem
-
-Approach this problem as if it is an application going to production. 
-We don't expect it to be perfect (no production code is), but we also don't want you to hack together a throw-away script. 
-This should be representative of something that you would be comfortable releasing to a production environment.  
-
-Also, spend whatever amount of time you think is reasonable.
-You may use whatever libraries, frameworks and tools that you think are appropriate, just provide any special setup instructions when you submit your solution.
-
-You will not get any extra point for solving the challenge fast. Speed of delivery is not what we are looking for in this challenge.
-
-We are looking for you to demonstrate your knowledge related to common software practices to include reusability, portability and encapsulation - to name a few.
-
-#### 4. Do not fork the question
-
-Please do not fork this repo.
-
-#### 5. No extra points for writing tests for the exercises
-
-If you want to write test for your solutions then please go ahead, but no points will be added for having tests, nor will points be deducted for not having any tests. 
-
-#### 6. Make commits incrementally
-
-Please do not squash commit before submission. We want to see incremental commits.
-
-#### 7. Commits after feedback should be incremental
-
-- Once you submit the answer you might get some feedback. 
-- If you change code based on feedback then please make those changes as separate commit. 
-- As mentioned above always make incremental commits.
-
-#### 8. There is no time limit and no extra point for doing it fast
-
-Most of the people taking the challenges already have a full time job.
-So take your time. Don't rush through it. There is no extra point
-for doing it fast.
-
-#### 9. Programming Language
-
-Your choice on programming language to use, however the Maxwell stack currently is Ruby and Javascript.
