--- todo/rego.md
@@ -1,7 +1,0 @@
-rego ./test/_.rb ./test/maxwell/_.rb ./lib/maxwell/_.rb ./lib/maxwell/interview/_.rb -- bundle exec rake test
-
-```
-rego ./**/*.rb ./**/**/*.rb ./**/**/**/*.rb ./**/**/**/*.rb -- bundle exec rake test
-
-
-```
